refactor: move getToc in create-pages.js
